requirements.txt
requests>=2.31.0
faiss-cpu>=1.10.0
numpy>=2.4.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
pypdf>=3.0.0
openai>=1.0.0
streamlit>=1.28.0



